Refactor shared processors
