+++ p-token/src/processor/burn_checked.rs
@@ -4,7 +4,7 @@ use pinocchio::{
     account_info::AccountInfo, program_error::ProgramError, pubkey::Pubkey, ProgramResult,
 };
 
-use super::burn::process_burn;
+use super::shared;
 
 #[inline(always)]
 pub fn process_burn_checked(
@@ -13,7 +13,7 @@ pub fn process_burn_checked(
     amount: u64,
     decimals: u8,
 ) -> ProgramResult {
-    process_burn(program_id, accounts, amount, Some(decimals))
+    shared::burn::process_burn(program_id, accounts, amount, Some(decimals))
 }
 
 pub struct BurnChecked<'a> {
