+++ p-token/src/processor/initialize_multisig2.rs
@@ -1,8 +1,8 @@
 use pinocchio::{account_info::AccountInfo, ProgramResult};
 
-use super::initialize_multisig::process_initialize_multisig;
+use super::shared;
 
 #[inline(always)]
 pub fn process_initialize_multisig2(accounts: &[AccountInfo], m: u8) -> ProgramResult {
-    process_initialize_multisig(accounts, m, false)
+    shared::initialize_multisig::process_initialize_multisig(accounts, m, false)
 }
