+++ p-token/src/processor/shared/toggle_account_state.rs
@@ -9,8 +9,9 @@ use token_interface::{
     },
 };
 
-use super::validate_owner;
+use crate::processor::validate_owner;
 
+#[inline(always)]
 pub fn process_toggle_account_state(
     program_id: &Pubkey,
     accounts: &[AccountInfo],
