+++ p-token/src/processor/transfer_checked.rs
@@ -4,7 +4,7 @@ use pinocchio::{
     account_info::AccountInfo, program_error::ProgramError, pubkey::Pubkey, ProgramResult,
 };
 
-use super::transfer::process_transfer;
+use super::shared;
 
 #[inline(always)]
 pub fn process_transfer_checked(
@@ -13,7 +13,7 @@ pub fn process_transfer_checked(
     amount: u64,
     decimals: u8,
 ) -> ProgramResult {
-    process_transfer(program_id, accounts, amount, Some(decimals))
+    shared::transfer::process_transfer(program_id, accounts, amount, Some(decimals))
 }
 
 pub struct TransferChecked<'a> {
